pyproject.toml
[tool.poetry]
name = "loupe"
version = "0.1.0"
description = ""
authors = ["alex lee <ajlee3141@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
spacy = "^3.0.0"
pandas = "^1.1.4"
numpy = "^1.19.4"
flask = "^1.1.2"
fastapi = "^0.62.0"
uvicorn = "^0.13.0"
python-multipart = "^0.0.5"
python-docx = "^0.8.10"
pdfplumber = "^0.5.25"
black = "^20.8b1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"